Remove dead code, fix int32 offset overflow, position_ids to int64
- Remove unused compute_freq_yarn, computeCosSinCacheKernel,
  compute_cos_sin_cache (cos_sin_cache is precomputed in Python)
- All buffer offsets (qkv, q_output, kv_cache, cos_sin_cache) changed
  from int to int64_t to prevent overflow on large models
  (GLM4 + H200: 356M KV slots overflows int32 at cacheSlot * 128)
- position_ids changed from int32 to int64 to match vLLM convention
- bench position_ids updated to int64

+++ bench/bench.py
@@ -81,7 +81,7 @@ def make_inputs(num_tokens, device="cuda", seed=42):
     kv_dim = NUM_HEADS_K * HEAD_DIM
 
     qkv = torch.randn(num_tokens, hidden, dtype=torch.bfloat16, device=device)
-    position_ids = (torch.arange(num_tokens, device=device) + 100).to(torch.int32)
+    position_ids = (torch.arange(num_tokens, device=device) + 100).to(torch.int64)
     q_weight = torch.randn(HEAD_DIM, dtype=torch.bfloat16, device=device) * 2.0
     k_weight = torch.randn(HEAD_DIM, dtype=torch.bfloat16, device=device) * 2.0
     q_output = torch.zeros(num_tokens, q_dim, dtype=torch.uint8, device=device)
